requirements.txt
astroid==2.4.2
blinker==1.4
click==7.1.2
Flask==1.1.2
Flask-JWT-Extended==3.24.1
Flask-Mail==0.9.1
flask-marshmallow==0.13.0
Flask-SQLAlchemy==2.4.4
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
marshmallow==3.7.1
mccabe==0.6.1
PyJWT==1.7.1
pylint==2.5.3
python-dotenv==0.14.0
six==1.15.0
SQLAlchemy==1.3.18
toml==0.10.1
typed-ast==1.4.1
Werkzeug==1.0.1
wrapt==1.12.1
